todo_app/requirements.txt
asgiref==3.5.2
Django==4.1
djangorestframework==3.13.1
psycopg2==2.9.3
PyJWT==1.7.1
pytz==2022.2.1
sqlparse==0.4.2
django-filter==22.1
